Simplify reach router component setup
